+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/govermentbody/pagemode/GovernmentBodyHeadcountPageModContentFactoryImpl.java
@@ -66,7 +66,7 @@ public final class GovernmentBodyHeadcountPageModContentFactoryImpl extends Abst
 			if (governmentBodyAnnualSummary != null) {
 				getGovernmentBodyMenuItemFactory().createGovernmentBodyMenuBar(menuBar, pageId,governmentBodyAnnualSummary.getName());
 				CardInfoRowUtil.createPageHeader(panel, panelContent,
-					GovernmentBodyTitleFormatter.formatTitle(governmentBodyAnnualSummary,
+					formatTitle(governmentBodyAnnualSummary,
 						GovernmentBodyDescriptionConstants.HEADCOUNT_HEADER),
 					GovernmentBodyDescriptionConstants.HEADCOUNT_SUBTITLE,
 					GovernmentBodyDescriptionConstants.HEADCOUNT_DESC);
