+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/govermentbody/pagemode/GovernmentBodyIncomePageModContentFactoryImpl.java
@@ -67,7 +67,7 @@ public final class GovernmentBodyIncomePageModContentFactoryImpl extends Abstrac
 			if (governmentBodyAnnualSummary != null) {
 				getGovernmentBodyMenuItemFactory().createGovernmentBodyMenuBar(menuBar, pageId,governmentBodyAnnualSummary.getName());
 				CardInfoRowUtil.createPageHeader(panel, panelContent,
-					GovernmentBodyTitleFormatter.formatTitle(governmentBodyAnnualSummary,
+					formatTitle(governmentBodyAnnualSummary,
 						GovernmentBodyDescriptionConstants.INCOME_HEADER),
 					GovernmentBodyDescriptionConstants.INCOME_SUBTITLE,
 					GovernmentBodyDescriptionConstants.INCOME_DESC);
