--- citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/govermentbody/pagemode/GovernmentBodyTitleFormatter.java
@@ -1,25 +1,0 @@
-package com.hack23.cia.web.impl.ui.application.views.user.govermentbody.pagemode;
-
-import java.util.Locale;
-
-import com.hack23.cia.service.external.esv.api.GovernmentBodyAnnualSummary;
-
-/**
- * The Class GovernmentBodyTitleFormatter.
- */
-public class GovernmentBodyTitleFormatter {
-
-    /**
-     * Format title.
-     *
-     * @param govBody the gov body
-     * @param pageTitle the page title
-     * @return the string
-     */
-    public static String formatTitle(final GovernmentBodyAnnualSummary govBody, final String pageTitle) {
-        if (govBody == null) {
-            return pageTitle;
-        }
-        return String.format(Locale.ENGLISH, "%s %s", pageTitle, govBody.getName());
-    }
-}
