+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/politician/pagemode/PoliticianDocumentActivityPageModContentFactoryImpl.java
@@ -63,7 +63,7 @@ public final class PoliticianDocumentActivityPageModContentFactoryImpl
 		getPoliticianMenuItemFactory().createPoliticianMenuBar(menuBar, pageId);
 
 		CardInfoRowUtil.createPageHeader(panel, panelContent,
-            PoliticianPageTitleFormatter.formatTitle(viewRiksdagenPolitician,
+            formatTitle(viewRiksdagenPolitician,
                 PoliticianPageTitleConstants.DOCUMENT_ACTIVITY_TITLE),
             PoliticianPageTitleConstants.ACTIVITY_SUBTITLE,
             PoliticianPageTitleConstants.DOC_ACTIVITY_DESC);
