+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/politician/pagemode/PoliticianIndicatorsPageModContentFactoryImpl.java
@@ -61,7 +61,7 @@ public final class PoliticianIndicatorsPageModContentFactoryImpl extends Abstrac
 		getPoliticianMenuItemFactory().createPoliticianMenuBar(menuBar, pageId);
 
 		CardInfoRowUtil.createPageHeader(panel, panelContent,
-            PoliticianPageTitleFormatter.formatTitle(viewRiksdagenPolitician,
+            formatTitle(viewRiksdagenPolitician,
                 PoliticianPageTitleConstants.INDICATORS_TITLE),
             PoliticianPageTitleConstants.INDICATOR_SUBTITLE,
             PoliticianPageTitleConstants.INDICATORS_DESC);
