+++ citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/politician/pagemode/PoliticianRoleListPageModContentFactoryImpl.java
@@ -61,7 +61,7 @@ public final class PoliticianRoleListPageModContentFactoryImpl extends AbstractP
 		getPoliticianMenuItemFactory().createPoliticianMenuBar(menuBar, pageId);
 
 		CardInfoRowUtil.createPageHeader(panel, panelContent,
-            PoliticianPageTitleFormatter.formatTitle(viewRiksdagenPolitician, PoliticianPageTitleConstants.ROLE_LIST_TITLE),
+            formatTitle(viewRiksdagenPolitician, PoliticianPageTitleConstants.ROLE_LIST_TITLE),
             PoliticianPageTitleConstants.ROLES_SUBTITLE,
             PoliticianPageTitleConstants.ROLE_LIST_DESC);
