--- citizen-intelligence-agency/src/main/java/com/hack23/cia/web/impl/ui/application/views/user/politician/pagemode/PoliticianTitleFormatter.java
@@ -1,21 +1,0 @@
-package com.hack23.cia.web.impl.ui.application.views.user.politician.pagemode;
-
-import com.hack23.cia.model.internal.application.data.politician.impl.ViewRiksdagenPolitician;
-
-/**
- * The Class PoliticianTitleFormatter.
- */
-public class PoliticianTitleFormatter {
-
-    /**
-     * Format title.
-     *
-     * @param politician the politician
-     * @return the string
-     */
-    public static String formatTitle(final ViewRiksdagenPolitician politician) {
-        return politician.getFirstName() + ' ' +
-               politician.getLastName() + '(' +
-               politician.getParty() + ')';
-    }
-}
